Refactoring backend to new namespace
