+++ services/core-api/app/api/incidents/namespace.py
@@ -5,7 +5,6 @@ from app.api.incidents.resources.mine_incident_status_codes import MineIncidentS
 from app.api.incidents.resources.mine_incident_followup_types import MineIncidentFollowupTypeResource
 from app.api.incidents.resources.mine_incident_determination_types import MineIncidentDeterminationTypeResource
 from app.api.incidents.resources.mine_incident_document_type_codes import MineIncidentDocumentTypeCodeResource
-from app.api.incidents.resources.mine_incident_notes import MineIncidentNoteListResource, MineIncidentNoteResource
 from app.api.incidents.resources.incidents_resource import IncidentsResource
 
 api = Namespace('incidents', description='Incidents actions/options')
@@ -16,6 +15,3 @@ api.add_resource(MineIncidentDeterminationTypeResource, '/determination-types')
 api.add_resource(MineIncidentStatusCodeResource, '/status-codes')
 api.add_resource(MineIncidentDocumentTypeCodeResource, '/document-types')
 api.add_resource(MineIncidentCategoryResource, '/category-codes')
-api.add_resource(MineIncidentNoteResource,
-                 '/<string:mine_incident_guid>/notes/<string:mine_incident_note_guid>')
-api.add_resource(MineIncidentNoteListResource, '/<string:mine_incident_guid>/notes')
